backend/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pymongo>=4.7.0
python-jose[cryptography]>=3.3.0
bcrypt>=4.1.0
pydantic>=2.7.0
pydantic-settings>=2.3.0
python-dotenv>=1.0.0
pytest>=8.2.0
httpx>=0.27.0
slowapi>=0.1.9
apscheduler>=3.10.4
python-multipart>=0.0.9
boto3>=1.34.0
python-json-logger>=2.0.7
cloudinary>=1.36.0
pillow>=10.1.0
reportlab>=4.0.7
openpyxl>=3.1.2
pandas>=2.1.4
numpy>=1.25.2
scikit-learn>=1.3.2
matplotlib==3.7.2
seaborn==0.12.2
websockets==12.0
redis==5.0.1
celery==5.3.4
flower==2.0.1
email-validator==2.1.0
aiofiles==23.2.1
jinja2==3.1.2
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
